Use "tap" for controlling updates
